cloudFunctions/requirements.txt
# Function dependencies, for example:
# package>=version
google-api-python-client
google-auth

